feat: add MH-Z19c and BME280 sensors with UART output and integrate into architecture

+++ Drivers_custom/sd/sd_manager.c
@@ -16,7 +16,7 @@ uint8_t BSP_SD_Init(void)
 }
 
 /* ==== READ ==== */
-uint8_t BSP_SD_ReadBlocks(uint32_t *pData,
+uint8_t BSP_SD_ReadBlocks(uint8_t *pData,
                           uint32_t BlockAddr,
                           uint32_t NumOfBlocks,
                           uint32_t Timeout)
@@ -36,7 +36,7 @@ uint8_t BSP_SD_ReadBlocks(uint32_t *pData,
 }
 
 /* ==== WRITE ==== */
-uint8_t BSP_SD_WriteBlocks(uint32_t *pData,
+uint8_t BSP_SD_WriteBlocks(uint8_t *pData,
                            uint32_t BlockAddr,
                            uint32_t NumOfBlocks,
                            uint32_t Timeout)
