+++ Drivers_custom/sd/sd_manager.h
@@ -13,12 +13,12 @@ extern "C" {
 
 uint8_t BSP_SD_Init(void);
 
-uint8_t BSP_SD_ReadBlocks(uint32_t *pData,
+uint8_t BSP_SD_ReadBlocks(uint8_t *pData,
                           uint32_t BlockAddr,
                           uint32_t NumOfBlocks,
                           uint32_t Timeout);
 
-uint8_t BSP_SD_WriteBlocks(uint32_t *pData,
+uint8_t BSP_SD_WriteBlocks(uint8_t *pData,
                            uint32_t BlockAddr,
                            uint32_t NumOfBlocks,
                            uint32_t Timeout);
